+++ apps/app/src/components/home/FollowedStores.tsx
@@ -1,12 +1,17 @@
-import { Dialog, SectionHeader, Spacer, TextButton } from '@habiti/components';
+import React from 'react';
+import { View, Pressable, StyleSheet } from 'react-native';
+import {
+	Avatar,
+	Dialog,
+	SectionHeader,
+	Spacer,
+	Typography
+} from '@habiti/components';
 import { useNavigation, NavigationProp } from '@react-navigation/native';
 import { FlashList } from '@shopify/flash-list';
-import React from 'react';
-import { View, StyleSheet } from 'react-native';
 
-import FollowedStoresItem from './FollowedStoresItem';
 import type { Store } from '../../data/types';
-import { HomeStackParamList } from '../../navigation/types';
+import type { HomeStackParamList } from '../../navigation/types';
 
 interface FollowedStoresProps {
 	followed: Store[];
@@ -58,7 +63,7 @@ const FollowedStoresMain: React.FC<FollowedStoresMainProps> = ({
 			horizontal
 			data={followed.slice(0, 8)}
 			keyExtractor={item => item.id}
-			contentContainerStyle={{ paddingLeft: 16 }}
+			contentContainerStyle={{ paddingLeft: 16, gap: 16 }}
 			renderItem={({ item }) => (
 				<FollowedStoresItem store={item} onPress={handleStorePress(item.id)} />
 			)}
@@ -66,9 +71,37 @@ const FollowedStoresMain: React.FC<FollowedStoresMainProps> = ({
 	);
 };
 
+interface FollowedStoresItemProps {
+	store: Store;
+	onPress(): void;
+}
+
+const FollowedStoresItem: React.FC<FollowedStoresItemProps> = ({
+	store,
+	onPress
+}) => (
+	<Pressable onPress={onPress}>
+		<Avatar
+			uri={store.image?.path}
+			size={68}
+			fallbackText={store.name}
+			circle
+		/>
+
+		<Spacer y={4} />
+
+		<Typography size='small' weight='medium' style={styles.name}>
+			{store.name}
+		</Typography>
+	</Pressable>
+);
+
 const styles = StyleSheet.create({
 	dialog: {
 		marginHorizontal: 16
+	},
+	name: {
+		textAlign: 'center'
 	}
 });
