+++ apps/app/src/components/store/CategorySelector.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
-import { StyleSheet, ScrollView } from 'react-native';
+import { View, Pressable, ScrollView, StyleSheet } from 'react-native';
+import { Spacer, Typography, useTheme } from '@habiti/components';
 
-import CategorySelectorItem from './CategorySelectorItem';
 import type { StoreProductCategory } from '../../data/types';
 
 interface CategorySelectorProps {
@@ -36,6 +36,40 @@ const CategorySelector: React.FC<CategorySelectorProps> = ({
 		))}
 	</ScrollView>
 );
+
+interface CategorySelectorItemProps {
+	name: string;
+	onPress(): void;
+	active: boolean;
+}
+
+const CategorySelectorItem: React.FC<CategorySelectorItemProps> = ({
+	name,
+	onPress,
+	active
+}) => {
+	const { theme } = useTheme();
+
+	return (
+		<Pressable onPress={onPress}>
+			<View style={{ paddingHorizontal: 16 }}>
+				<Typography weight='medium' variant={active ? 'primary' : 'secondary'}>
+					{name}
+				</Typography>
+			</View>
+			<Spacer y={8} />
+			<View style={{ alignItems: 'center' }}>
+				<View
+					style={{
+						height: 2,
+						backgroundColor: active ? theme.text.primary : 'transparent',
+						width: '100%'
+					}}
+				/>
+			</View>
+		</Pressable>
+	);
+};
 
 const styles = StyleSheet.create({
 	container: {
