--- apps/app/src/components/store/StoreListItem.tsx
@@ -1,45 +1,0 @@
-import React from 'react';
-import { StyleSheet, Pressable } from 'react-native';
-import { formatNaira } from '@habiti/common';
-import { CustomImage, Spacer, Typography } from '@habiti/components';
-
-import type { Product } from '../../data/types';
-
-interface StoreListItemProps {
-	item: Product;
-	onPress(): void;
-	side: 'left' | 'right';
-}
-
-const StoreListItem: React.FC<StoreListItemProps> = ({
-	item,
-	onPress,
-	side
-}) => (
-	<Pressable
-		key={item.id}
-		style={[
-			styles.pressable,
-			{ [side === 'left' ? 'marginLeft' : 'marginRight']: 16 }
-		]}
-		onPress={onPress}
-	>
-		<CustomImage height={200} style={styles.image} uri={item.images[0]?.path} />
-		<Typography weight='medium'>{item.name}</Typography>
-		<Spacer y={2} />
-		<Typography variant='secondary'>{formatNaira(item.unitPrice)}</Typography>
-	</Pressable>
-);
-
-const styles = StyleSheet.create({
-	pressable: {
-		flex: 1,
-		margin: 8
-	},
-	image: {
-		width: '100%',
-		marginBottom: 8
-	}
-});
-
-export default StoreListItem;
