+++ apps/app/src/screens/Settings.tsx
@@ -1,10 +1,15 @@
-import { Button, ScrollableScreen } from '@habiti/components';
+import React from 'react';
+import { View, StyleSheet, Pressable } from 'react-native';
+import {
+	Button,
+	ScrollableScreen,
+	Icon,
+	Typography,
+	useTheme
+} from '@habiti/components';
 import { useNavigation } from '@react-navigation/native';
 import { StackNavigationProp } from '@react-navigation/stack';
-import React from 'react';
-import { StyleSheet } from 'react-native';
 
-import SettingRow from '../components/settings/SettingRow';
 import { useShallow } from 'zustand/react/shallow';
 import useStore from '../state';
 import { ThemeMap } from '../utils/theme';
@@ -38,9 +43,59 @@ const Settings: React.FC = () => {
 	);
 };
 
+interface SettingRowProps {
+	name: string;
+	onPress(): void;
+	displayValue?: string;
+}
+
+const SettingRow: React.FC<SettingRowProps> = ({
+	name,
+	onPress,
+	displayValue
+}) => {
+	const { theme } = useTheme();
+
+	return (
+		<Pressable
+			style={[styles.container, { borderBottomColor: theme.border.color }]}
+			onPress={onPress}
+		>
+			<Typography>{name}</Typography>
+			<View style={styles.right}>
+				{displayValue ? (
+					<Typography variant='secondary'>{displayValue}</Typography>
+				) : null}
+				<Icon
+					name='chevron-right'
+					color={theme.text.secondary}
+					size={24}
+					style={styles.icon}
+				/>
+			</View>
+		</Pressable>
+	);
+};
+
 const styles = StyleSheet.create({
 	logOut: {
 		margin: 8
+	},
+	container: {
+		width: '100%',
+		height: 45,
+		paddingHorizontal: 16,
+		borderBottomWidth: StyleSheet.hairlineWidth,
+		flexDirection: 'row',
+		justifyContent: 'space-between',
+		alignItems: 'center'
+	},
+	right: {
+		flexDirection: 'row',
+		alignItems: 'center'
+	},
+	icon: {
+		marginLeft: 8
 	}
 });
